+++ vck/src/commonMain/kotlin/at/asitplus/wallet/lib/iso/ByteStringWrapperItemsRequestSerializer.kt
@@ -1,0 +1,29 @@
+package at.asitplus.wallet.lib.iso
+
+import at.asitplus.signum.indispensable.cosef.io.ByteStringWrapper
+import kotlinx.serialization.KSerializer
+import kotlinx.serialization.builtins.ByteArraySerializer
+import kotlinx.serialization.decodeFromByteArray
+import kotlinx.serialization.descriptors.PrimitiveKind
+import kotlinx.serialization.descriptors.PrimitiveSerialDescriptor
+import kotlinx.serialization.descriptors.SerialDescriptor
+import kotlinx.serialization.encodeToByteArray
+import kotlinx.serialization.encoding.Decoder
+import kotlinx.serialization.encoding.Encoder
+
+object ByteStringWrapperItemsRequestSerializer : KSerializer<ByteStringWrapper<ItemsRequest>> {
+
+    override val descriptor: SerialDescriptor =
+        PrimitiveSerialDescriptor("ByteStringWrapperItemsRequestSerializer", PrimitiveKind.STRING)
+
+    override fun serialize(encoder: Encoder, value: ByteStringWrapper<ItemsRequest>) {
+        val bytes = vckCborSerializer.encodeToByteArray(value.value)
+        encoder.encodeSerializableValue(ByteArraySerializer(), bytes)
+    }
+
+    override fun deserialize(decoder: Decoder): ByteStringWrapper<ItemsRequest> {
+        val bytes = decoder.decodeSerializableValue(ByteArraySerializer())
+        return ByteStringWrapper(vckCborSerializer.decodeFromByteArray(bytes), bytes)
+    }
+
+}
