+++ vck/src/commonMain/kotlin/at/asitplus/wallet/lib/iso/DeviceAuth.kt
@@ -1,0 +1,16 @@
+package at.asitplus.wallet.lib.iso
+
+import at.asitplus.signum.indispensable.cosef.CoseSigned
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+/**
+ * Part of the ISO/IEC 18013-5:2021 standard: Data structure for mdoc request (8.3.2.1.2.1)
+ */
+@Serializable
+data class DeviceAuth(
+    @SerialName("deviceSignature")
+    val deviceSignature: CoseSigned? = null,
+    @SerialName("deviceMac")
+    val deviceMac: CoseSigned? = null, // TODO is COSE_Mac0
+)
