+++ vck/src/commonMain/kotlin/at/asitplus/wallet/lib/iso/DeviceRequest.kt
@@ -3,19 +3,7 @@
 package at.asitplus.wallet.lib.iso
 
 import at.asitplus.KmmResult.Companion.wrap
-import at.asitplus.signum.indispensable.cosef.CoseSigned
-import at.asitplus.signum.indispensable.cosef.io.ByteStringWrapper
-import io.matthewnelson.encoding.base16.Base16
-import io.matthewnelson.encoding.core.Encoder.Companion.encodeToString
 import kotlinx.serialization.*
-import kotlinx.serialization.builtins.ByteArraySerializer
-import kotlinx.serialization.builtins.MapSerializer
-import kotlinx.serialization.builtins.serializer
-import kotlinx.serialization.cbor.ByteString
-import kotlinx.serialization.cbor.ValueTags
-import kotlinx.serialization.descriptors.*
-import kotlinx.serialization.encoding.*
-import okio.ByteString.Companion.toByteString
 
 /**
  * Part of the ISO/IEC 18013-5:2021 standard: Data structure for mdoc request (8.3.2.1.2.1)
@@ -53,521 +41,3 @@ data class DeviceRequest(
     }
 }
 
-/**
- * Part of the ISO/IEC 18013-5:2021 standard: Data structure for mdoc request (8.3.2.1.2.1)
- */
-@Serializable
-data class DocRequest(
-    @SerialName("itemsRequest")
-    @Serializable(with = ByteStringWrapperItemsRequestSerializer::class)
-    @ValueTags(24U)
-    val itemsRequest: ByteStringWrapper<ItemsRequest>,
-    @SerialName("readerAuth")
-    val readerAuth: CoseSigned? = null,
-) {
-    override fun toString(): String {
-        return "DocRequest(itemsRequest=${itemsRequest.value}, readerAuth=$readerAuth)"
-    }
-
-}
-
-/**
- * Part of the ISO/IEC 18013-5:2021 standard: Data structure for mdoc request (8.3.2.1.2.1)
- */
-@Serializable
-data class ItemsRequest(
-    @SerialName("docType")
-    val docType: String,
-    @SerialName("nameSpaces")
-    val namespaces: Map<String, ItemsRequestList>,
-    @SerialName("requestInfo")
-    val requestInfo: Map<String, String>? = null,
-)
-
-
-/**
- * Convenience class with a custom serializer ([ItemsRequestListSerializer]) to prevent
- * usage of the type `Map<String, Map<String, Boolean>>` in [ItemsRequest.namespaces].
- */
-@Serializable(with = ItemsRequestListSerializer::class)
-data class ItemsRequestList(
-    val entries: List<SingleItemsRequest>
-)
-
-/**
- * Convenience class with a custom serializer ([ItemsRequestListSerializer]) to prevent
- * usage of the type `Map<String, Map<String, Boolean>>` in [ItemsRequest.namespaces].
- */
-data class SingleItemsRequest(
-    val key: String,
-    val value: Boolean,
-)
-
-/**
- * Serializes [ItemsRequestList.entries] as an "inline map",
- * having [SingleItemsRequest.key] as the map key and [SingleItemsRequest.value] as the map value,
- * for the map represented by [ItemsRequestList].
- */
-object ItemsRequestListSerializer : KSerializer<ItemsRequestList> {
-
-    override val descriptor: SerialDescriptor = mapSerialDescriptor(
-        keyDescriptor = PrimitiveSerialDescriptor("key", PrimitiveKind.INT),
-        valueDescriptor = listSerialDescriptor<Byte>(),
-    )
-
-    override fun serialize(encoder: Encoder, value: ItemsRequestList) {
-        encoder.encodeStructure(descriptor) {
-            var index = 0
-            value.entries.forEach {
-                this.encodeStringElement(descriptor, index++, it.key)
-                this.encodeBooleanElement(descriptor, index++, it.value)
-            }
-        }
-    }
-
-    override fun deserialize(decoder: Decoder): ItemsRequestList {
-        val entries = mutableListOf<SingleItemsRequest>()
-        decoder.decodeStructure(descriptor) {
-            lateinit var key: String
-            var value: Boolean
-            while (true) {
-                val index = decodeElementIndex(descriptor)
-                if (index == CompositeDecoder.DECODE_DONE) {
-                    break
-                } else if (index % 2 == 0) {
-                    key = decodeStringElement(descriptor, index)
-                } else if (index % 2 == 1) {
-                    value = decodeBooleanElement(descriptor, index)
-                    entries += SingleItemsRequest(key, value)
-                }
-            }
-        }
-        return ItemsRequestList(entries)
-    }
-}
-
-
-/**
- * Part of the ISO/IEC 18013-5:2021 standard: Data structure for mdoc request (8.3.2.1.2.1)
- */
-@Serializable
-data class DeviceResponse(
-    @SerialName("version")
-    val version: String,
-    @SerialName("documents")
-    val documents: Array<Document>? = null,
-    @SerialName("documentErrors")
-    val documentErrors: Array<Pair<String, UInt>>? = null,
-    @SerialName("status")
-    val status: UInt,
-) {
-    fun serialize() = vckCborSerializer.encodeToByteArray(this)
-
-    override fun equals(other: Any?): Boolean {
-        if (this === other) return true
-        if (other == null || this::class != other::class) return false
-
-        other as DeviceResponse
-
-        if (version != other.version) return false
-        if (documents != null) {
-            if (other.documents == null) return false
-            if (!documents.contentEquals(other.documents)) return false
-        } else if (other.documents != null) return false
-        if (documentErrors != null) {
-            if (other.documentErrors == null) return false
-            if (!documentErrors.contentEquals(other.documentErrors)) return false
-        } else if (other.documentErrors != null) return false
-        return status == other.status
-    }
-
-    override fun hashCode(): Int {
-        var result = version.hashCode()
-        result = 31 * result + (documents?.contentHashCode() ?: 0)
-        result = 31 * result + (documentErrors?.contentHashCode() ?: 0)
-        result = 31 * result + status.hashCode()
-        return result
-    }
-
-    companion object {
-        fun deserialize(it: ByteArray) = kotlin.runCatching {
-            vckCborSerializer.decodeFromByteArray<DeviceResponse>(it)
-        }.wrap()
-    }
-}
-
-/**
- * Part of the ISO/IEC 18013-5:2021 standard: Data structure for mdoc request (8.3.2.1.2.1)
- */
-@Serializable
-data class Document(
-    @SerialName("docType")
-    val docType: String, // this is relevant for deserializing the elementValue of IssuerSignedItem
-    @SerialName("issuerSigned")
-    val issuerSigned: IssuerSigned,
-    @SerialName("deviceSigned")
-    val deviceSigned: DeviceSigned,
-    @SerialName("errors")
-    val errors: Map<String, Map<String, Int>>? = null,
-) {
-
-    fun serialize() = vckCborSerializer.encodeToByteArray(this)
-    override fun equals(other: Any?): Boolean {
-        if (this === other) return true
-        if (other !is Document) return false
-
-        if (docType != other.docType) return false
-        if (issuerSigned != other.issuerSigned) return false
-        if (deviceSigned != other.deviceSigned) return false
-        if (errors != other.errors) return false
-
-        return true
-    }
-
-    override fun hashCode(): Int {
-        var result = docType.hashCode()
-        result = 31 * result + issuerSigned.hashCode()
-        result = 31 * result + deviceSigned.hashCode()
-        result = 31 * result + (errors?.hashCode() ?: 0)
-        return result
-    }
-
-    companion object {
-        fun deserialize(it: ByteArray) = kotlin.runCatching {
-            vckCborSerializer.decodeFromByteArray<Document>(it)
-        }.wrap()
-    }
-}
-
-
-/**
- * Part of the ISO/IEC 18013-5:2021 standard: Data structure for mdoc request (8.3.2.1.2.1)
- */
-@Serializable
-data class IssuerSigned private constructor(
-    @SerialName("nameSpaces")
-    @Serializable(with = NamespacedIssuerSignedListSerializer::class)
-    val namespaces: Map<String, @Contextual IssuerSignedList>? = null,
-    @SerialName("issuerAuth")
-    val issuerAuth: CoseSigned,
-) {
-
-    constructor(
-        namespacedItems: Map<String, List<IssuerSignedItem>>,
-        issuerAuth: CoseSigned,
-        tag: Byte? = null
-    ) : this(issuerAuth = issuerAuth, namespaces = namespacedItems.map { (ns, value) ->
-        ns to IssuerSignedList(
-            value.map { item ->
-                ByteStringWrapper(
-                    item,
-                    item.serialize(ns).let { tag?.let { tg -> it.wrapInCborTag(tg) } ?: it })
-            })
-    }.toMap())
-
-
-    fun getIssuerAuthPayloadAsMso() = issuerAuth.payload?.stripCborTag(24)
-        ?.let { vckCborSerializer.decodeFromByteArray(ByteStringWrapperMobileSecurityObjectSerializer, it).value }
-
-    fun serialize() = vckCborSerializer.encodeToByteArray(this)
-    override fun equals(other: Any?): Boolean {
-        if (this === other) return true
-        if (other !is IssuerSigned) return false
-
-        if (issuerAuth != other.issuerAuth) return false
-        if (namespaces != other.namespaces) return false
-
-        return true
-    }
-
-    override fun hashCode(): Int {
-        var result = issuerAuth.hashCode()
-        result = 31 * result + (namespaces?.hashCode() ?: 0)
-        return result
-    }
-
-    companion object {
-        fun deserialize(it: ByteArray) = kotlin.runCatching {
-            vckCborSerializer.decodeFromByteArray<IssuerSigned>(it)
-        }.wrap()
-    }
-}
-
-object NamespacedIssuerSignedListSerializer : KSerializer<Map<String, IssuerSignedList>> {
-
-    private val mapSerializer = MapSerializer(String.serializer(), object : IssuerSignedListSerializer("") {})
-
-    override val descriptor = mapSerializer.descriptor
-
-    override fun deserialize(decoder: Decoder): Map<String, IssuerSignedList> = NamespacedMapEntryDeserializer().let {
-        MapSerializer(it.namespaceSerializer, it.itemSerializer).deserialize(decoder)
-    }
-
-    class NamespacedMapEntryDeserializer {
-        lateinit var key: String
-
-        val namespaceSerializer = NamespaceSerializer()
-        val itemSerializer = IssuerSignedListSerializer()
-
-        inner class NamespaceSerializer internal constructor() : KSerializer<String> {
-            override val descriptor = PrimitiveSerialDescriptor("ISO namespace", PrimitiveKind.STRING)
-
-            override fun deserialize(decoder: Decoder): String = decoder.decodeString().apply { key = this }
-
-            override fun serialize(encoder: Encoder, value: String) {
-                encoder.encodeString(value).also { key = value }
-            }
-
-        }
-
-        inner class IssuerSignedListSerializer internal constructor() : KSerializer<IssuerSignedList> {
-            override val descriptor = mapSerializer.descriptor
-
-            override fun deserialize(decoder: Decoder): IssuerSignedList =
-                decoder.decodeSerializableValue(IssuerSignedListSerializer(key))
-
-
-            override fun serialize(encoder: Encoder, value: IssuerSignedList) {
-                encoder.encodeSerializableValue(IssuerSignedListSerializer(key), value)
-            }
-
-        }
-    }
-
-
-    override fun serialize(encoder: Encoder, value: Map<String, IssuerSignedList>) {
-        mapSerializer.serialize(encoder, value)
-    }
-
-}
-
-
-/**
- * Convenience class with a custom serializer ([IssuerSignedListSerializer]) to prevent
- * usage of the type `Map<String, List<ByteStringWrapper<IssuerSignedItem>>>` in [namespaces].
- */
-data class IssuerSignedList(
-    val entries: List<ByteStringWrapper<IssuerSignedItem>>
-) {
-    override fun toString(): String {
-        return "IssuerSignedList(entries=${entries.map { it.value }})"
-    }
-
-    override fun equals(other: Any?): Boolean {
-        if (this === other) return true
-        if (other !is IssuerSignedList) return false
-
-        if (entries != other.entries) return false
-
-        return true
-    }
-
-    override fun hashCode(): Int {
-        return 31 * entries.hashCode()
-    }
-}
-
-/**
- * Serializes [IssuerSignedList.entries] as an "inline list",
- * having serialized instances of [IssuerSignedItem] as the values.
- */
-open class IssuerSignedListSerializer(private val namespace: String) : KSerializer<IssuerSignedList> {
-
-    override val descriptor: SerialDescriptor = object : SerialDescriptor {
-        @ExperimentalSerializationApi
-        override val elementsCount: Int = 1
-
-        @ExperimentalSerializationApi
-        override val kind: SerialKind = StructureKind.LIST
-
-        @ExperimentalSerializationApi
-        override val serialName: String = "kotlin.collections.ArrayList"
-
-        @ExperimentalSerializationApi
-        override fun getElementAnnotations(index: Int): List<Annotation> {
-            @OptIn(ExperimentalUnsignedTypes::class)
-            return listOf(ValueTags(24U))
-        }
-
-        @ExperimentalSerializationApi
-        override fun getElementDescriptor(index: Int): SerialDescriptor {
-            return Byte.serializer().descriptor
-        }
-
-        @ExperimentalSerializationApi
-        override fun getElementIndex(name: String): Int {
-            return name.toInt()
-        }
-
-        @ExperimentalSerializationApi
-        override fun getElementName(index: Int): String {
-            return index.toString()
-        }
-
-        @ExperimentalSerializationApi
-        override fun isElementOptional(index: Int): Boolean {
-            return false
-        }
-    }
-
-
-    override fun serialize(encoder: Encoder, value: IssuerSignedList) {
-        var index = 0
-        encoder.encodeCollection(descriptor, value.entries.size) {
-            value.entries.forEach {
-                encodeSerializableElement(descriptor, index++, ByteArraySerializer(), it.value.serialize(namespace))
-            }
-        }
-    }
-
-    override fun deserialize(decoder: Decoder): IssuerSignedList {
-        val entries = mutableListOf<ByteStringWrapper<IssuerSignedItem>>()
-        decoder.decodeStructure(descriptor) {
-            while (true) {
-                val index = decodeElementIndex(descriptor)
-                if (index == CompositeDecoder.DECODE_DONE) {
-                    break
-                }
-                val readBytes = decoder.decodeSerializableValue(ByteArraySerializer())
-                val issuerSignedItem = IssuerSignedItem.deserialize(readBytes, namespace).getOrThrow()
-                entries += ByteStringWrapper(issuerSignedItem, readBytes)
-            }
-        }
-        return IssuerSignedList(entries)
-    }
-}
-
-/**
- * Part of the ISO/IEC 18013-5:2021 standard: Data structure for mdoc request (8.3.2.1.2.1)
- */
-
-data class IssuerSignedItem(
-    @SerialName(PROP_DIGEST_ID)
-    val digestId: UInt,
-    @SerialName(PROP_RANDOM)
-    @ByteString
-    val random: ByteArray,
-    @SerialName(PROP_ELEMENT_ID)
-    val elementIdentifier: String,
-    @SerialName(PROP_ELEMENT_VALUE)
-    val elementValue: Any,
-) {
-
-
-    fun serialize(namespace: String) = vckCborSerializer.encodeToByteArray(IssuerSignedItemSerializer(namespace), this)
-
-    override fun equals(other: Any?): Boolean {
-        if (this === other) return true
-        if (other == null || this::class != other::class) return false
-
-        other as IssuerSignedItem
-
-        if (digestId != other.digestId) return false
-        if (!random.contentEquals(other.random)) return false
-        if (elementIdentifier != other.elementIdentifier) return false
-        if (elementValue is ByteArray && other.elementValue is ByteArray) return elementValue.contentEquals(other.elementValue)
-        if (elementValue is IntArray && other.elementValue is IntArray) return elementValue.contentEquals(other.elementValue)
-        if (elementValue is BooleanArray && other.elementValue is BooleanArray) return elementValue.contentEquals(other.elementValue)
-        if (elementValue is CharArray && other.elementValue is CharArray) return elementValue.contentEquals(other.elementValue)
-        if (elementValue is ShortArray && other.elementValue is ShortArray) return elementValue.contentEquals(other.elementValue)
-        if (elementValue is LongArray && other.elementValue is LongArray) return elementValue.contentEquals(other.elementValue)
-        if (elementValue is FloatArray && other.elementValue is FloatArray) return elementValue.contentEquals(other.elementValue)
-        if (elementValue is DoubleArray && other.elementValue is DoubleArray) return elementValue.contentEquals(other.elementValue)
-        return if (elementValue is Array<*> && other.elementValue is Array<*>) elementValue.contentDeepEquals(other.elementValue)
-        //It was time for Thomas to leave. He had seen everything.
-        else elementValue == other.elementValue
-    }
-
-    override fun hashCode(): Int {
-        var result = digestId.hashCode()
-        result = 31 * result + random.contentHashCode()
-        result = 31 * result + elementIdentifier.hashCode()
-        result = 31 * result + elementValue.hashCode()
-        return result
-    }
-
-    override fun toString(): String {
-        return "IssuerSignedItem(digestId=$digestId," +
-                " random=${random.encodeToString(Base16(strict = true))}," +
-                " elementIdentifier='$elementIdentifier'," +
-                " elementValue=$elementValue)"
-    }
-
-    companion object {
-        fun deserialize(it: ByteArray, namespace: String) = kotlin.runCatching {
-            vckCborSerializer.decodeFromByteArray(IssuerSignedItemSerializer(namespace), it)
-        }.wrap()
-
-        internal const val PROP_DIGEST_ID = "digestID"
-        internal const val PROP_RANDOM = "random"
-        internal const val PROP_ELEMENT_ID = "elementIdentifier"
-        internal const val PROP_ELEMENT_VALUE = "elementValue"
-    }
-}
-
-
-/**
- * Part of the ISO/IEC 18013-5:2021 standard: Data structure for mdoc request (8.3.2.1.2.1)
- */
-@Serializable
-data class DeviceSigned(
-    @SerialName("nameSpaces")
-    @ByteString
-    @ValueTags(24U)
-    val namespaces: ByteArray,
-    @SerialName("deviceAuth")
-    val deviceAuth: DeviceAuth,
-) {
-    override fun equals(other: Any?): Boolean {
-        if (this === other) return true
-        if (other == null || this::class != other::class) return false
-
-        other as DeviceSigned
-
-        if (!namespaces.contentEquals(other.namespaces)) return false
-        return deviceAuth == other.deviceAuth
-    }
-
-    override fun hashCode(): Int {
-        var result = namespaces.contentHashCode()
-        result = 31 * result + deviceAuth.hashCode()
-        return result
-    }
-
-}
-
-
-/**
- * Part of the ISO/IEC 18013-5:2021 standard: Data structure for mdoc request (8.3.2.1.2.1)
- */
-@Serializable
-data class DeviceAuth(
-    @SerialName("deviceSignature")
-    val deviceSignature: CoseSigned? = null,
-    @SerialName("deviceMac")
-    val deviceMac: CoseSigned? = null, // TODO is COSE_Mac0
-)
-
-
-object ByteStringWrapperItemsRequestSerializer : KSerializer<ByteStringWrapper<ItemsRequest>> {
-
-    override val descriptor: SerialDescriptor =
-        PrimitiveSerialDescriptor("ByteStringWrapperItemsRequestSerializer", PrimitiveKind.STRING)
-
-    override fun serialize(encoder: Encoder, value: ByteStringWrapper<ItemsRequest>) {
-        val bytes = vckCborSerializer.encodeToByteArray(value.value)
-        encoder.encodeSerializableValue(ByteArraySerializer(), bytes)
-    }
-
-    override fun deserialize(decoder: Decoder): ByteStringWrapper<ItemsRequest> {
-        val bytes = decoder.decodeSerializableValue(ByteArraySerializer())
-        return ByteStringWrapper(vckCborSerializer.decodeFromByteArray(bytes), bytes)
-    }
-
-}
-
-fun ByteArray.stripCborTag(tag: Byte) = this.dropWhile { it == 0xd8.toByte() }.dropWhile { it == tag }.toByteArray()
-
-fun ByteArray.wrapInCborTag(tag: Byte) = byteArrayOf(0xd8.toByte()) + byteArrayOf(tag) + this
-
-fun ByteArray.sha256(): ByteArray = toByteString().sha256().toByteArray()
