+++ vck/src/commonMain/kotlin/at/asitplus/wallet/lib/iso/DeviceSigned.kt
@@ -1,0 +1,37 @@
+package at.asitplus.wallet.lib.iso
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+import kotlinx.serialization.cbor.ByteString
+import kotlinx.serialization.cbor.ValueTags
+
+/**
+ * Part of the ISO/IEC 18013-5:2021 standard: Data structure for mdoc request (8.3.2.1.2.1)
+ */
+@OptIn(ExperimentalUnsignedTypes::class)
+@Serializable
+data class DeviceSigned(
+    @SerialName("nameSpaces")
+    @ByteString
+    @ValueTags(24U)
+    val namespaces: ByteArray,
+    @SerialName("deviceAuth")
+    val deviceAuth: DeviceAuth,
+) {
+    override fun equals(other: Any?): Boolean {
+        if (this === other) return true
+        if (other == null || this::class != other::class) return false
+
+        other as DeviceSigned
+
+        if (!namespaces.contentEquals(other.namespaces)) return false
+        return deviceAuth == other.deviceAuth
+    }
+
+    override fun hashCode(): Int {
+        var result = namespaces.contentHashCode()
+        result = 31 * result + deviceAuth.hashCode()
+        return result
+    }
+
+}
