+++ vck/src/commonMain/kotlin/at/asitplus/wallet/lib/iso/IssuerSignedList.kt
@@ -1,0 +1,28 @@
+package at.asitplus.wallet.lib.iso
+
+import at.asitplus.signum.indispensable.cosef.io.ByteStringWrapper
+
+/**
+ * Convenience class with a custom serializer ([IssuerSignedListSerializer]) to prevent
+ * usage of the type `Map<String, List<ByteStringWrapper<IssuerSignedItem>>>` in [IssuerSigned.namespaces].
+ */
+data class IssuerSignedList(
+    val entries: List<ByteStringWrapper<IssuerSignedItem>>
+) {
+    override fun toString(): String {
+        return "IssuerSignedList(entries=${entries.map { it.value }})"
+    }
+
+    override fun equals(other: Any?): Boolean {
+        if (this === other) return true
+        if (other !is IssuerSignedList) return false
+
+        if (entries != other.entries) return false
+
+        return true
+    }
+
+    override fun hashCode(): Int {
+        return 31 * entries.hashCode()
+    }
+}
