+++ vck/src/commonMain/kotlin/at/asitplus/wallet/lib/iso/ItemsRequest.kt
@@ -1,0 +1,17 @@
+package at.asitplus.wallet.lib.iso
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+/**
+ * Part of the ISO/IEC 18013-5:2021 standard: Data structure for mdoc request (8.3.2.1.2.1)
+ */
+@Serializable
+data class ItemsRequest(
+    @SerialName("docType")
+    val docType: String,
+    @SerialName("nameSpaces")
+    val namespaces: Map<String, ItemsRequestList>,
+    @SerialName("requestInfo")
+    val requestInfo: Map<String, String>? = null,
+)
