+++ vck/src/commonMain/kotlin/at/asitplus/wallet/lib/iso/ItemsRequestList.kt
@@ -1,0 +1,12 @@
+package at.asitplus.wallet.lib.iso
+
+import kotlinx.serialization.Serializable
+
+/**
+ * Convenience class with a custom serializer ([ItemsRequestListSerializer]) to prevent
+ * usage of the type `Map<String, Map<String, Boolean>>` in [ItemsRequest.namespaces].
+ */
+@Serializable(with = ItemsRequestListSerializer::class)
+data class ItemsRequestList(
+    val entries: List<SingleItemsRequest>
+)
