+++ vck/src/commonMain/kotlin/at/asitplus/wallet/lib/iso/ItemsRequestListSerializer.kt
@@ -1,0 +1,48 @@
+package at.asitplus.wallet.lib.iso
+
+import kotlinx.serialization.KSerializer
+import kotlinx.serialization.descriptors.*
+import kotlinx.serialization.encoding.*
+
+/**
+ * Serializes [ItemsRequestList.entries] as an "inline map",
+ * having [SingleItemsRequest.key] as the map key and [SingleItemsRequest.value] as the map value,
+ * for the map represented by [ItemsRequestList].
+ */
+object ItemsRequestListSerializer : KSerializer<ItemsRequestList> {
+
+    override val descriptor: SerialDescriptor = mapSerialDescriptor(
+        keyDescriptor = PrimitiveSerialDescriptor("key", PrimitiveKind.INT),
+        valueDescriptor = listSerialDescriptor<Byte>(),
+    )
+
+    override fun serialize(encoder: Encoder, value: ItemsRequestList) {
+        encoder.encodeStructure(descriptor) {
+            var index = 0
+            value.entries.forEach {
+                this.encodeStringElement(descriptor, index++, it.key)
+                this.encodeBooleanElement(descriptor, index++, it.value)
+            }
+        }
+    }
+
+    override fun deserialize(decoder: Decoder): ItemsRequestList {
+        val entries = mutableListOf<SingleItemsRequest>()
+        decoder.decodeStructure(descriptor) {
+            lateinit var key: String
+            var value: Boolean
+            while (true) {
+                val index = decodeElementIndex(descriptor)
+                if (index == CompositeDecoder.DECODE_DONE) {
+                    break
+                } else if (index % 2 == 0) {
+                    key = decodeStringElement(descriptor, index)
+                } else if (index % 2 == 1) {
+                    value = decodeBooleanElement(descriptor, index)
+                    entries += SingleItemsRequest(key, value)
+                }
+            }
+        }
+        return ItemsRequestList(entries)
+    }
+}
