+++ vck/src/commonMain/kotlin/at/asitplus/wallet/lib/iso/SingleItemsRequest.kt
@@ -1,0 +1,10 @@
+package at.asitplus.wallet.lib.iso
+
+/**
+ * Convenience class with a custom serializer ([ItemsRequestListSerializer]) to prevent
+ * usage of the type `Map<String, Map<String, Boolean>>` in [ItemsRequest.namespaces].
+ */
+data class SingleItemsRequest(
+    val key: String,
+    val value: Boolean,
+)
